refactor(web): route submission through SessionDriver planSubmission/startTurn
sendMessage becomes shared choreography only: busy-gate planning, queueing,
mode switch, conflict retries, and the error path. The three dispatch
submission paths (submit / append-steer / continue with idempotent retry
ids) move verbatim into DispatchSessionDriver; the local dialog-turn path
(optimistic turn, state machine, worktree materialization, model sync, ACP
fork, backend-retry fallback) moves into LocalSessionDriver.

Load-bearing invariants preserved and now stated in the interface:
- planSubmission tests steer-eligibility before queueing, so a steerable
  dispatch message can never be parked in a queue that only drains for
  state-machine sessions (new regression test).
- ensureReady may return void synchronously; a projection adds its
  optimistic turn with no microtask boundary after the user's send action
  (pinned by the existing projects-immediately test).
- The dispatch_pending_<jobId> turn id, the submit-snapshot cursor CAS
  arguments, and the per-session retry-id lifecycle are byte-identical.
- Steered/continued messages still skip the shared post-submission
  bookkeeping ('detached' outcome).

Idempotent retry ids generalize into session-drivers/idempotency.ts;
syncSessionModelSelection moves to utils/modelSync (re-exported).
