+++ src/web-ui/src/flow_chat/session-drivers/local/LocalSessionDriver.ts
@@ -8,23 +8,41 @@
  */
 
 import { agentAPI } from '@/infrastructure/api/service-api/AgentAPI';
+import { ACPClientAPI } from '@/infrastructure/api/service-api/ACPClientAPI';
 import { sessionAPI } from '@/infrastructure/api/service-api/SessionAPI';
+import { worktreeAPI } from '@/infrastructure/api/service-api/WorktreeAPI';
+import { globalEventBus } from '@/infrastructure/event-bus';
 import { createLogger } from '@/shared/utils/logger';
 import { stateMachineManager } from '../../state-machine';
 import { SessionExecutionEvent, SessionExecutionState } from '../../state-machine/types';
-import type { FlowChatContext, SessionConfig } from '../../services/flow-chat-manager/types';
-import type { SessionCascadeRemoval, SessionCreationSeed, SessionDriver } from '../types';
+import type { FlowChatContext, SessionConfig, DialogTurn } from '../../services/flow-chat-manager/types';
+import type {
+  SessionCascadeRemoval,
+  SessionCreationSeed,
+  SessionDriver,
+  StartTurnInput,
+  StartTurnResult,
+  SubmissionPlan,
+  TurnTracker,
+} from '../types';
+import {
+  FLOWCHAT_PIN_TURN_TO_TOP_EVENT,
+  type FlowChatPinTurnToTopRequest,
+} from '../../events/flowchatNavigation';
 import {
   getModelMaxTokens,
   resolveModelForSessionCreation,
 } from '../../utils/modelResolution';
+import { syncSessionModelSelection } from '../../utils/modelSync';
 import { markCurrentTurnItemsAsCancelled } from '../../utils/turnCancellation';
 import {
   requireSessionProjectWorkspacePath,
   sessionProjectWorkspacePath,
 } from '../../utils/sessionWorkspace';
+import { sessionWorktreeMaterializationPlan } from '../../utils/sessionWorktree';
 import { cleanupSaveState } from '../../services/flow-chat-manager/PersistenceModule';
 import { cleanupSessionBuffers } from '../../services/flow-chat-manager/TextChunkModule';
+import { applyGeneratingTitlePlaceholder } from '../shared';
 
 const log = createLogger('LocalSessionDriver');
 
@@ -194,5 +212,210 @@ export const localSessionDriver: SessionDriver = {
     }
 
     return success;
+  },
+
+  planSubmission(): SubmissionPlan {
+    // Local sessions park messages while busy; the drain listener replays
+    // them when the state machine returns to IDLE.
+    return { kind: 'queue' };
+  },
+
+  async steer(): Promise<void> {
+    throw new Error('Local sessions queue messages while a turn is running');
+  },
+
+  async startTurn(
+    context: FlowChatContext,
+    input: StartTurnInput,
+    tracker: TurnTracker,
+  ): Promise<StartTurnResult> {
+    const {
+      sessionId,
+      message,
+      displayMessage,
+      currentAgentType,
+      acpClientId,
+      isFirstMessage,
+      readySession,
+      options,
+    } = input;
+
+    const dialogTurnId = options?.turnId?.trim() ||
+      `dialog_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
+    const hasImages = (options?.imageContexts?.length ?? 0) > 0;
+
+    const dialogTurn: DialogTurn = {
+      id: dialogTurnId,
+      sessionId: sessionId,
+      agentType: currentAgentType,
+      userMessage: {
+        id: `user_${Date.now()}`,
+        content: displayMessage || message,
+        timestamp: Date.now(),
+        hasImages,
+        images: options?.imageDisplayData,
+        metadata: options?.userMessageMetadata,
+      },
+      modelRounds: [],
+      // Images are attached for multimodal primary models or reduced to text placeholders for text-only models.
+      // We don't run a separate frontend "image pre-analysis" phase here.
+      status: 'pending',
+      startTime: Date.now()
+    };
+
+    context.flowChatStore.addDialogTurn(sessionId, dialogTurn);
+    tracker.createdLocalTurnId = dialogTurnId;
+    const pinRequest: FlowChatPinTurnToTopRequest = {
+      sessionId,
+      turnId: dialogTurnId,
+      behavior: 'auto',
+      source: 'send-message',
+      pinMode: 'sticky-latest',
+    };
+    globalEventBus.emit(FLOWCHAT_PIN_TURN_TO_TOP_EVENT, pinRequest, 'MessageModule');
+
+    const isRestoringHistoricalSession =
+      readySession.isHistorical || context.pendingHistoryLoads.has(sessionId);
+    if (isRestoringHistoricalSession) {
+      context.processingManager.clearSessionStatus(sessionId);
+      context.flowChatStore.deleteDialogTurn(sessionId, dialogTurnId);
+      throw new Error('Session history is still restoring, please retry once loading finishes');
+    }
+
+    const startOk = await stateMachineManager.transition(sessionId, SessionExecutionEvent.START, {
+      taskId: sessionId,
+      dialogTurnId,
+    });
+    if (!startOk) {
+      const currentState = stateMachineManager.getCurrentState(sessionId);
+      throw new Error(`Session is still busy finishing the previous turn (current state: ${currentState})`);
+    }
+
+    context.processingManager.registerStatus({
+      sessionId: sessionId,
+      status: 'thinking',
+      message: '',
+      metadata: { sessionId: sessionId, dialogTurnId }
+    });
+
+    if (readySession.config.worktreeIsolationRequested !== undefined) {
+      const materialization = sessionWorktreeMaterializationPlan(readySession);
+      if (materialization) {
+        log.info('Materializing requested worktree after prompt submission', {
+          sessionId,
+          enabled: materialization.enabled,
+          projectWorkspacePath: materialization.projectWorkspacePath,
+        });
+        const result = await worktreeAPI.bindSession(
+          sessionId,
+          materialization.enabled,
+          globalThis.crypto?.randomUUID?.() ?? `worktree-first-turn-${Date.now()}`,
+          materialization.projectWorkspacePath,
+        );
+        context.flowChatStore.updateSessionExecutionTarget(sessionId, {
+          workspacePath: result.workspacePath,
+          projectWorkspacePath: result.projectWorkspacePath,
+          workspaceId: result.workspaceId,
+          executionTarget: result.executionTarget,
+        });
+        if (result.retainedWorktreePath) {
+          log.warn('Released worktree retained because it contains local work', {
+            sessionId,
+            retainedWorktreePath: result.retainedWorktreePath,
+          });
+        }
+      }
+      context.flowChatStore.setSessionWorktreeIsolationRequested(sessionId, undefined);
+    }
+
+    if (isFirstMessage) {
+      applyGeneratingTitlePlaceholder(context, sessionId, message);
+    }
+
+    if (!acpClientId) {
+      await syncSessionModelSelection(context, sessionId, currentAgentType);
+    }
+
+    const updatedSession = context.flowChatStore.getState().sessions.get(sessionId);
+    if (!updatedSession) {
+      throw new Error(`Session lost after adding dialog turn: ${sessionId}`);
+    }
+
+    context.contentBuffers.set(sessionId, new Map());
+    context.activeTextItems.set(sessionId, new Map());
+
+    const workspacePath = updatedSession.workspacePath;
+    const projectWorkspacePath = sessionProjectWorkspacePath(updatedSession);
+
+    if (acpClientId) {
+      await ACPClientAPI.startDialogTurn({
+        sessionId,
+        clientId: acpClientId,
+        userInput: message,
+        originalUserInput: displayMessage || message,
+        turnId: dialogTurnId,
+        workspacePath,
+        imageContexts: options?.imageContexts,
+        userMessageMetadata: options?.userMessageMetadata,
+        remoteConnectionId: updatedSession.remoteConnectionId,
+        remoteSshHost: updatedSession.remoteSshHost,
+      });
+      context.flowChatStore.updateSessionLastSubmittedMode(sessionId, currentAgentType);
+    } else {
+      try {
+        await agentAPI.startDialogTurn({
+          sessionId: sessionId,
+          userInput: message,
+          originalUserInput: displayMessage || message,
+          turnId: dialogTurnId,
+          agentType: currentAgentType,
+          workspacePath,
+          projectWorkspacePath,
+          remoteConnectionId: updatedSession.remoteConnectionId,
+          remoteSshHost: updatedSession.remoteSshHost,
+          imageContexts: options?.imageContexts,
+          userMessageMetadata: options?.userMessageMetadata,
+          execution: options?.execution,
+        });
+        context.flowChatStore.updateSessionLastSubmittedMode(sessionId, currentAgentType);
+      } catch (error: any) {
+        if (error?.message?.includes('Session does not exist') || error?.message?.includes('Not found')) {
+          log.warn('Backend session still not found, retrying creation', {
+            sessionId: sessionId,
+            dialogTurnsCount: updatedSession.dialogTurns.length
+          });
+
+          // Lazy import: SessionModule routes lifecycle calls through the
+          // driver registry, so a static import would create a module cycle.
+          const { retryCreateBackendSession } =
+            await import('../../services/flow-chat-manager/SessionModule');
+          await retryCreateBackendSession(context, sessionId);
+
+          await agentAPI.startDialogTurn({
+            sessionId: sessionId,
+            userInput: message,
+            originalUserInput: displayMessage || message,
+            turnId: dialogTurnId,
+            agentType: currentAgentType,
+            workspacePath,
+            projectWorkspacePath,
+            remoteConnectionId: updatedSession.remoteConnectionId,
+            remoteSshHost: updatedSession.remoteSshHost,
+            imageContexts: options?.imageContexts,
+            userMessageMetadata: options?.userMessageMetadata,
+            execution: options?.execution,
+          });
+          context.flowChatStore.updateSessionLastSubmittedMode(sessionId, currentAgentType);
+        } else {
+          throw error;
+        }
+      }
+    }
+
+    const sessionStateMachine = stateMachineManager.get(sessionId);
+    if (sessionStateMachine) {
+      sessionStateMachine.getContext().taskId = sessionId;
+    }
+    return 'completed';
   },
 };
